requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pypdf==5.0.1
openai==1.51.2
python-multipart==0.0.9